refactor: hooks 폴더 절대 경로 수정
