+++ frontend/src/hooks/authentication/index.ts
@@ -1,1 +1,1 @@
-export { useAuth } from './useAuth';
+export { useAuth } from '@/hooks/authentication/useAuth';
