+++ frontend/src/hooks/category/index.ts
@@ -1,2 +1,2 @@
-export { useCategory } from './useCategory';
-export { useCategoryNameValidation } from './useCategoryNameValidation';
+export { useCategory } from '@/hooks/category/useCategory';
+export { useCategoryNameValidation } from '@/hooks/category/useCategoryNameValidation';
