+++ frontend/src/hooks/template/index.ts
@@ -1,2 +1,2 @@
-export { useSourceCode } from './useSourceCode';
-export { useTag } from './useTag';
+export { useSourceCode } from '@/hooks/template/useSourceCode';
+export { useTag } from '@/hooks/template/useTag';
